requirements.txt
fastapi>=0.100.0,<1.0
uvicorn[standard]==0.22.0
SQLAlchemy>=1.4,<3.0
alembic==1.11.1
pydantic>=2.0,<3.0
PyYAML==6.0.3
python-multipart==0.0.6
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0
pytest-asyncio==0.22.0